health bar UI working
The script is basic, but there’s a lot of potential as to where to go.
Tomorrow morning I’ll build the health animation where it will shrink
and grow as needed, and then connect it to the player character to test
while I begin working on the boss script.