bossClass made, firing purple bullets
So there’s a couple things we need
1) A proper boss Sprite (Just using the skeleton right now)
2) To rotate the bullets properly
3) To have the bullets interact with push / pull

Once we do those things, I can start implementing patterns and such to
make it feel like a real boss fight. Once the bullets can be fired back
I can link my health bar to the boss (it is on the player rn for
testing) and we can get the ball rolling.