Ammo now firing from Player, Though Needs to be Tuned
Also, I was thinking about changing the boss mechanic somewhat?

Since the boss is going to be a cat with a handgun, we can add a
western theme to him, And make it so that our character cannot move
forward / back but only side to side. Like one of the old galaga games
or something, we could also have a cool feeling where u have to
constantly be dodging bullets, although we would have to build bullet
patterns around that idea to make it still challenging. Anyways, here’s
the ammo firing.